Broadcast messages.  Moved sockets to own file. Add username. overflow scroll
