+++ hangman-chat/server/views/chatScreen.pug
@@ -4,12 +4,12 @@ include mixins/_letterBoard
 
 block content
 	
-	div.modal
-		form(action="" class="form--modal")
+	div#modal.modal.modal--active
+		form(id="form-modal"  action="" class="form--modal")
 			h2 Hangman chat
 			label(for="username") What should we call you?
 			fieldset
-				input(id="username" type="text")
+				input(id="input-username" type="text" autofocus)
 				input(type="submit" value="👉")
 
 	main
@@ -17,11 +17,11 @@ block content
 			+letterBoard()
 			+hangmanWords()
 
-		section
+		section.message--con
 			ul#messages
 
 			form(id="chat-form" class="chat--form" action="")
-				input(id="m" type="text" autocomplete="on")
+				input(id="message" type="text" autocomplete="on")
 				input(type="submit" value="👉")
 
 	script(type="text/javascript" src="/socket.io/socket.io.js")
